pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "style-shift"
version = "0.1.0"
description = "Neural style transfer tool based on AdaIN"
authors = [
    {name = "StyleShift Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "Pillow>=9.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.990",
    "gradio>=3.0.0",
]

[project.scripts]
style-shift = "style_shift.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["style_shift*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=style_shift --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
